Eliminate remaining duplication: Consolidate AccountValidator methods, extract TransferCreationHelper, refactor ControllerErrorHandler
